Update Dashboard to be compatible with r1145.
Autoconverted from SVN (revision:1156)

+++ src/dashboard/src/main/models.py
@@ -13,7 +13,7 @@ class Job(models.Model):
   jobuuid = models.CharField(max_length=150, primary_key=True, db_column='jobUUID')
   jobtype = models.CharField(max_length=750, db_column='jobType', blank=True)
   createdtime = models.DateTimeField(db_column='createdTime')
-  createdtimedec = models.IntegerField(null=True, db_column='createdTimeDec', blank=True)
+  createdtimedec = models.DecimalField(null=True, db_column='createdTimeDec', blank=True)
   directory = models.CharField(max_length=750, blank=True)
   sipuuid = models.CharField(max_length=150, db_column='SIPUUID', blank=True)
   currentstep = models.CharField(max_length=150, db_column='currentStep', blank=True)
